requirements.txt
asgiref==3.4.1
Django==3.2.6
django-etc==1.3.0
django-froala-editor==3.2.2
django-hitcount==1.3.3
gunicorn==20.1.0
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
